refactor: rename & move dom.js file
Rename
- dom.js -> dom-selector.js

Move to modules folder
